ERC20 balance implementation broken into util files
