+++ src/blockchains/erc20/erc20_worker.ts
@@ -130,6 +130,9 @@ export class ERC20Worker extends BaseWorker {
     if ('extract_exact_overwrite' === this.settings.CONTRACT_MODE) {
       if (this.allOldContracts.length > 0) {
         events = await this.getPastEventsFun(this.web3Wrapper, interval.fromBlock, interval.toBlock, this.allOldContracts, timestampsCache);
+        if (this.settings.EXTEND_TRANSFERS_WITH_BALANCES && interval.fromBlock > this.settings.MULTICALL_DEPLOY_BLOCK) {
+          await extendTransfersWithBalances((this.web3Wrapper as Web3Wrapper).getWeb3(), events);
+        }
         changeContractAddresses(events, this.contractsOverwriteArray);
       }
 
@@ -137,6 +140,9 @@ export class ERC20Worker extends BaseWorker {
         const rawEvents = await this.getPastEventsFun(this.web3Wrapper, interval.fromBlock, interval.toBlock, this.contractsUnmodified,
           timestampsCache);
 
+        if (this.settings.EXTEND_TRANSFERS_WITH_BALANCES && interval.fromBlock > this.settings.MULTICALL_DEPLOY_BLOCK) {
+          await extendTransfersWithBalances((this.web3Wrapper as Web3Wrapper).getWeb3(), events);
+        }
         for (const event of rawEvents) {
           events.push(event);
         }
