+++ src/blockchains/erc20/lib/add_balances.ts
@@ -1,102 +1,18 @@
 import { assertIsDefinedLazy } from '../../../lib/utils';
 import { ERC20Transfer } from '../erc20_types';
 import { Web3 } from 'web3';
-import { ZERO_ADDRESS } from './fetch_events';
+import { MULTICALL_ABI, ERC20_ABI, MULTICALL_ADDRESS } from './abis';
+import * as Utils from './balance_utils';
 
-type AddressContractToBalance = Map<string, string>;
-type BlockNumberToBalances = Map<number, AddressContractToBalance>;
-type AddressContract = [string, string]
-type BlockNumberAddressContractBalance = [number, string, string, string]
-type BlockNumberToAffectedAddresses = Map<number, ValueSet>;
 
-const MAX_BALANCES_PER_QUERY = 50;
+type BlockNumberToBalances = Map<number, Utils.AddressContractToBalance>;
+type BlockNumberToAffectedAddresses = Map<number, Utils.ValueSet>;
+
 const MULTICALL_FAILURE = "multicall_failure"
 
-const multicallAbi = [
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "internalType": "bool",
-        "name": "requireSuccess",
-        "type": "bool"
-      },
-      {
-        "components": [
-          {
-            "internalType": "address",
-            "name": "target",
-            "type": "address"
-          },
-          {
-            "internalType": "bytes",
-            "name": "callData",
-            "type": "bytes"
-          }
-        ],
-        "internalType": "struct Multicall3.Call[]",
-        "name": "calls",
-        "type": "tuple[]"
-      }
-    ],
-    "name": "tryAggregate",
-    "outputs": [
-      {
-        "components": [
-          {
-            "internalType": "bool",
-            "name": "success",
-            "type": "bool"
-          },
-          {
-            "internalType": "bytes",
-            "name": "returnData",
-            "type": "bytes"
-          }
-        ],
-        "internalType": "struct Multicall3.Result[]",
-        "name": "returnData",
-        "type": "tuple[]"
-      }
-    ],
-    "stateMutability": "payable",
-    "type": "function"
-  }
-];
-
-const multicallAddress = '0xcA11bde05977b3631167028862bE2a173976CA11';
-
-const ERC20_ABI = [
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "_owner",
-        "type": "address"
-      }
-    ],
-    "name": "balanceOf",
-    "outputs": [
-      {
-        "name": "balance",
-        "type": "uint256"
-      }
-    ],
-    "type": "function"
-  }
-];
-
-function decodeRevertReason(web3: Web3, errorData: string) {
-  try {
-    return web3.eth.abi.decodeParameter('string', '0x' + errorData.slice(10));
-  } catch (e) {
-    return 'Unable to decode error data';
-  }
-}
-
-async function getBalancesPerBlock(web3: Web3, addressContract: AddressContract[], blockNumber: number)
-  : Promise<BlockNumberAddressContractBalance[]> {
-  const multicall = new web3.eth.Contract(multicallAbi, multicallAddress);
+async function getBalancesPerBlock(web3: Web3, addressContract: Utils.AddressContract[], blockNumber: number)
+  : Promise<Utils.BlockNumberAddressContractBalance[]> {
+  const multicall = new web3.eth.Contract(MULTICALL_ABI, MULTICALL_ADDRESS);
 
   const calls = addressContract.map(([address, contractAddress]) => {
     const contract = new web3.eth.Contract(ERC20_ABI, contractAddress);
@@ -116,7 +32,7 @@ async function getBalancesPerBlock(web3: Web3, addressContract: AddressContract[
     console.error('Error while calling multicall:', error);
 
     if (error.cause.data) {
-      const decodedReason = decodeRevertReason(web3, error.cause.data);
+      const decodedReason = Utils.decodeRevertReason(web3, error.cause.data);
       console.error('Revert reason:', decodedReason);
     }
     throw error;
@@ -128,7 +44,7 @@ async function getBalancesPerBlock(web3: Web3, addressContract: AddressContract[
 
   let index = 0;
 
-  const result: BlockNumberAddressContractBalance[] = [];
+  const result: Utils.BlockNumberAddressContractBalance[] = [];
   addressContract.forEach(([address, contractAddress]) => {
     const resultMulticall = resultsMulticall[index];
 
@@ -145,74 +61,35 @@ async function getBalancesPerBlock(web3: Web3, addressContract: AddressContract[
   return result;
 }
 
-function addToSet(set: ValueSet, event: ERC20Transfer, web3: Web3) {
-
-  if (isAddressEligableForBalance(web3, event.from, event.contract)) {
-    set.add([event.from, event.contract]);
-  }
-  if (isAddressEligableForBalance(web3, event.to, event.contract)) {
-    set.add([event.to, event.contract]);
-  }
-}
-
-function concatAddressAndContract(address: string, contract: string): string {
-  return address + "-" + contract
-}
-
-function addToMap(map: AddressContractToBalance, balance: BlockNumberAddressContractBalance) {
-  //console.log(`Adding balance ${balance[0]}-${balance[1]}-${balance[2]}-${balance[3]}`)
-  const concatenation = concatAddressAndContract(balance[1], balance[2])
-  map.set(concatenation, balance[3])
-}
-
-function breakNeededBalancesPerBatch(input: AddressContract[]): AddressContract[][] {
-  const result = [];
-  for (let i = 0; i < input.length; i += MAX_BALANCES_PER_QUERY) {
-    result.push(input.slice(i, i + MAX_BALANCES_PER_QUERY));
-  }
-  return result;
-}
-
-function isAddressEligableForBalance(web3: Web3, address: string, contract: string): boolean {
-  // TODO migrate deprecation
-  return web3.utils.isAddress(address) && address !== ZERO_ADDRESS && address !== contract
-}
 
 export async function extendTransfersWithBalances(web3: Web3, events: ERC20Transfer[]) {
   // Identify for which addresses we need to fetch balances
   const neededBalances: BlockNumberToAffectedAddresses = events.reduce((acc, event) => {
     const blockNumberSet = acc.get(event.blockNumber)
     if (blockNumberSet) {
-      addToSet(blockNumberSet, event, web3);
+      Utils.addToSet(blockNumberSet, event);
     }
     else {
-      const newSet = new ValueSet();
-      addToSet(newSet, event, web3);
+      const newSet = new Utils.ValueSet();
+      Utils.addToSet(newSet, event);
       acc.set(event.blockNumber, newSet)
     }
     return acc;
   }, new Map() as BlockNumberToAffectedAddresses)
 
 
-  // const promises: Promise<AddressContractBalance[]>[] = Array.from(neededBalances.entries()).flatMap(([blockNumber, addressSet]) => {
-  //   return breakNeededBalancesPerBatch(addressSet.values()).map((addressContract: AddressContract[]) => {
-  //     return getBalancesPerBlock(web3, addressContract, blockNumber)
-  //   })
-
-  // })
-
-  const addresses: [number, AddressContract[]][] = [];
+  const addresses: [number, Utils.AddressContract[]][] = [];
 
   for (const [blockNumber, addressSet] of Array.from(neededBalances.entries())) {
-    const brokenPerBatchAddresses: AddressContract[][] = breakNeededBalancesPerBatch(addressSet.values())
-    brokenPerBatchAddresses.map((addressContract: AddressContract[]) => {
+    const brokenPerBatchAddresses: Utils.AddressContract[][] = Utils.breakNeededBalancesPerBatch(addressSet.values())
+    brokenPerBatchAddresses.map((addressContract: Utils.AddressContract[]) => {
       addresses.push([blockNumber, addressContract])
     })
   }
 
-  const results: BlockNumberAddressContractBalance[] = []
+  const results: Utils.BlockNumberAddressContractBalance[] = []
   for (const blockNumberAddress of addresses) {
-    let result: BlockNumberAddressContractBalance[] = [];
+    let result: Utils.BlockNumberAddressContractBalance[] = [];
     try {
       result = await getBalancesPerBlock(web3, blockNumberAddress[1], blockNumberAddress[0])
     }
@@ -233,11 +110,11 @@ export async function extendTransfersWithBalances(web3: Web3, events: ERC20Trans
     const blockNumber = result[0];
     const balances = acc.get(blockNumber)
     if (balances) {
-      addToMap(balances, result);
+      Utils.addToMap(balances, result);
     }
     else {
-      const newMap: AddressContractToBalance = new Map();
-      addToMap(newMap, result);
+      const newMap: Utils.AddressContractToBalance = new Map();
+      Utils.addToMap(newMap, result);
       acc.set(blockNumber, newMap)
     }
     ++index;
@@ -250,49 +127,17 @@ export async function extendTransfersWithBalances(web3: Web3, events: ERC20Trans
 
 
   for (const event of events) {
-    if (isAddressEligableForBalance(web3, event.from, event.contract)) {
-      event.fromBalance = balanceMap.get(event.blockNumber)?.get(concatAddressAndContract(event.from, event.contract))
+    if (Utils.isAddressEligableForBalance(event.from, event.contract)) {
+      event.fromBalance = balanceMap.get(event.blockNumber)?.get(Utils.concatAddressAndContract(event.from, event.contract))
       assertIsDefinedLazy(event.fromBalance, () => `'from' balance should have been resovled for event ${JSON.stringify(event)}`)
     }
 
-    if (isAddressEligableForBalance(web3, event.to, event.contract)) {
-      event.toBalance = balanceMap.get(event.blockNumber)?.get(concatAddressAndContract(event.to, event.contract))
+    if (Utils.isAddressEligableForBalance(event.to, event.contract)) {
+      event.toBalance = balanceMap.get(event.blockNumber)?.get(Utils.concatAddressAndContract(event.to, event.contract))
       assertIsDefinedLazy(event.toBalance, () => `'to' balance should have been resovled for event ${JSON.stringify(event)}`)
     }
   }
 }
 
-class ValueSet {
-  private elements: AddressContract[];
 
-  constructor() {
-    this.elements = [];
-  }
-
-  has(item: AddressContract) {
-    return this.elements.some(element => this.isEqual(element, item));
-  }
-
-  add(item: AddressContract) {
-    if (!this.has(item)) {
-      this.elements.push(item);
-    }
-  }
-
-  isEqual(array1: AddressContract, array2: AddressContract) {
-    if (array1.length !== array2.length) {
-      return false;
-    }
-    for (let i = 0; i < array1.length; i++) {
-      if (array1[i] !== array2[i]) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  values() {
-    return [...this.elements];
-  }
-}
